requirements.txt
SpeechRecognition
beautifulsoup4
python-vlc
youtube-dl
pyowm
wikipedia
gTTS
pyttsx3